Fix symbol required in clamav_append_log interface
The label of clamd's log files is clamd_var_log_t instead of
clamav_log_t, which in fact has not been defined by clamav.pp. If such
un-decleared symbol is required by the clamav_append_log interface, the
whole optional block that calls this interface would not be enabled as
expected.

Signed-off-by: Harry Ciao <qingtao.cao at windriver.com>