Mark the modules directory as semanage_store_t
Previously, type transitions (on file/dir) occurred when semanage_t tried to create a directory inside a
selinux_config_t. Recently, this has changed so that this only occurs when said directory (that is created) is called
"modules".

However, on existing systems, this directory already exists (with the selinux_config_t type). As there is no file
context to say otherwise, loading a policy fails (as the "tmp/" dir that it creates doesn't transition) and restoring
the contexts of /etc/selinux recursively doesn't help either.

By adding the definition for the "modules/" directory, we should now be able to run a restorecon -R /etc/selinux once
and have the policy be loaded correctly again.