Add in transitions for queue types when the queues are created
At startup, postfix creates the missing queue directories inside
/var/spool/postfix. This is done by the master process. However, since the
/var/spool/postfix directory is labeled postfix_spool_t and there were no named
file transitions, all created queues became postfix_spool_t. This meant that an
administrator had to relabel the directories afterwards.

This patch adds in the necessary named file transitions for those directories,
with create and setattr (for changing ownership of the directories) for the
postfix master domain.

Also include the fowner capability, needed for when postfix chown's the queues.