Put in a file transition for regular tmp files as well
Commit d9d9f2b8c5 introduced the skype_tmp_t domain for the temporary files and
socket files created by the application. However, it only put in a file
transition for the sock_file class. As a result, skype_t wants to create a
regular tmp_t file (rather than a skype_tmp_t file) which of course it doesn't
hold the necessary rights for.

Add in the type transition for regular files as well, so that it creates a
skype_tmp_t one instead of tmp_t.