Add in administrative interface(s) for puppet
As the puppet domain has an named init script (puppet_initrc_exec_t and
puppetmaster_initrc_exec_t) we provide administrative interfaces to be used by
the users that need to admin the puppet domains.

Also assign this to the sysadm_t/sysadm_r user for general system administration
support.