Have syslog-ng.persist label survive relabeling
Syslog-ng already holds a filetrans when it writes files in the var_lib_t
resources, causing the files to be labeled syslogd_var_lib_t.

One of these files is the /var/lib/misc/syslog-ng.persist file. However, because
no file context was defined for this file, a system-wide relabeling caused it to
be turned back into var_lib_t.

This fixes the context back to syslogd_var_lib_t.