Initial dirmngr policy module
Ported from Debian with changes

Dirmngr is a server for managing and downloading certificate revocation
lists (CRLs) for X.509 certificates and for downloading the certificates
themselves. Dirmngr also handles OCSP requests as an alternative to
CRLs. Dirmngr is either invoked internally by gpgsm (from GnuPG 2) or
when running as a system daemon through the dirmngr-client tool.

Signed-off-by: Dominick Grift <dominick.grift@gmail.com>