Keep file contexts local
Unlike what is used in refpolicy, I think it is much cleaner if file contexts
related to the application the module reflects are within the file context file.
This does mean that "elsewhere" defined types must be accepted in the context,
but as long as these types are part of the base install (or as a depending
module without optional_policy() statement) this should be okay.

One main advantage to this - beyond clarity - is that the contexts file on a
users' system will not contain paths for files that are of applications he
doesn't have.

Doing this for a few shorewall contexts for now, will update as these come
along.