System logger can write to logfiles
The system logger, like syslog-ng, should be able to write to log files it has
to manage. As this does not include only the var_log_t labeled ones, we allow
write (and setattr) rights on the logfile files, and add_entry rights onto the
directories that are labeled as logfile too.

See bug #440128