Fixing cron access
Update the system logger policy to not allow write access on all possible log
files (too many log types are marked as logfile even though many of them are not
meant to be managed by the system logger). Instead, we use specific rights - in
this case for the cron log file.

Introduce a named file transition for when the system logger creates the initial
cron.log file. Also allow the write, setattr and create rights. We don't
implement full _manage_ rights as system loggers are not meant to remove log
files (unless someone points me to a use case for that).