Postfix creates defer and deferred queue locations
At startup, the Postfix daemon will check if the defer and deferred queues are
available. If not, it will create them. Introduce the proper file tansitions to
support this.

Also include the fowner capability as the postfix_master_t domain now requires
this for the initial setup of the queues.