Allow sandbox to log violations
When the sandbox (running in portage_sandbox_t) detects a violation, it will try
to log this into /var/log:sandbox. However, the portage_sandbox_t domain
currently is not allowed to do anything with this logs. As a result, the
violations are not logged.

Allow the portage_sandbox_t domain to generate logs (as portage_log_t) as well
as clean them up (sandbox will remove older violation logs if the process id of
the current violation would result in an existing log file to be overwritten).