Include documentation for generated tunables
When tunables are generated through a template, their documentation is not added
to the XML file that is parsed by "semanage boolean". As the current method for
obtaining the information is not M4-aware, we cannot get the information from
the policy sources themselves.

As a result, we now add in a gentoo_tunables.xml file that contains static
declarations of the booleans.