Move 389-ds-base provided dirsrv module to repository
The 389-ds-base package provides policies as part of the package. By moving this
policy into our own policy repository, we can manage this as we do with the
others. It will also simplify the maintenance of the ebuild and SELinux-related
activities.

Without this action, the ebuild fails to load the policy properly for each and
every supported policy type (the "targeted" policy is hardcoded and loading in
the policy doesn't properly select the store to use.

See also bug #445484