Updates on stunnel policy
Allow the stunnel init script to read the stunnel configuration (needed to find
out where the PID file will be).

Also allow stunnel to read generic certificates (for instance when providing the
CA certificate for mutual authentication purposes).

Allow stunnel to bind on any port (name_bind privilege) as the primary purpose
of stunnel is to provide encrypted tunneling regardless of the underlying
service.