Move out at code from cron
This is a first attempt to move out the at code from the cron domain towards its
own domain. The services cron and at are different in how they act, plus that
atd is not SELinux-aware, so I think we cannot use the default_contexts approach.

Also, at interprets the job files but doesn't execute it (it spawns a shell that
reads in the file and executes the content).

Still a lot of work, but it is fun to see how all this interacts...

What I will *not* do is extend atd_t (or if I move to an atd_userjob_t) with
more privileges.