Puppet module helper scripts are puppet_var_lib_t
As the module helper scripts are puppet_var_lib_t, we allow puppet_t to execute
these files. Another method would be to have the scripts marked bin_t, but as
these paths are depending on module names this does not seem feasible at first
sight. More investigation is needed on this matter though.