Fix bug #453724 - chsh fails
Running chsh fails (error on writing to /etc/passwd).

Analysis of the denials (cfr bug report) tell us that we need to provide chfn_t
(the domain in which chsh runs) read rights on the file context definitions, and
furthermore also execute rights on nscd_t to flush the cache.