Support passwordless service handling
When using Gentoo init scripts, we transition to run_init_t before calling
run_init. That means that when run_init runs and checks the PAM services, the
pam_rootok.so calls will look for the previous context (which already is
run_init_t) and check if that context is allowed to "accept root" as a
sufficient authentication measure.

run_init_t doesn't (well, didn't with this patch), which makes this check fail.
When calling services with run_init directly (like "run_init rc-service sshd
status") then the previous context is (most likely) the sysadm_t one, which does
have this permission.

See also
http://blog.siphos.be/2013/04/not-needing-run_init-for-password-less-service-management/