Fix bug 463006 - Support shared file access for pulseaudio
The pulseaudio setup uses shared files in /dev/shm where all pulseaudio-capable
domains should have the proper access to. The policy already supports marking
the tmpfs file types as pulseaudio_tmpfsfile's, but this wasn't set everywhere
(the bug report mentions mplayer, mozilla, mozilla_plugin, thunderbird).

As this is similar to ALSA, I decided to use the same approach as it is, imo,
quite manageable:

- created a pulseaudio_client_domain() interface that takes two arguments
  (1.) the client domain itself (which will be marked through attribute
       pulseaudio_client)
  (2.) the tmpfs type (which will be marked through attribute
       pulseaudio_tmpfsfile)
- give pulseaudio_client_domain() on the given types

With that done, support for all pulseaudio-related matters are automatically
assigned the moment that the pulseaudio module is loaded.