Partial fix bug 463690 - Kernel creates device-mapper
In the initramfs, processes run as kernel_t. When vgscan runs and creates device
nodes, it also creates /dev/device-mapper. This block device should be of type
fixed_disk_device_t and it currently is (was) device_t.

Add in a storage_dev_filetrans_fixed_disk call for this specific device.