Enable python-exec support
In Gentoo, /usr/sbin/semanage is a symlink to /usr/bin/python-exec which will
then decide which Python version to execute. As semanage is only labeled as
semanage_exec_t if it is a regular file, it now remains bin_t.

We have two choices here - either relabel /usr/sbin/semanage itself, or the
semanage-python* files. We pick the second, because we don't know what
rights/permissions python-exec needs, but it shouldn't be semanage_t.