Revert back to apache_manage_all_user_content
Previously, apache_manage_all_user_content was deprecated and potentially moved
to apache_manage_all_content. However, this was a wrong step - the interface was
put back and its content was updated to do what it was supposed to do.

Hence we revert the change to apache_manage_all_content back to
apache_manage_all_user_content.