Adding files_read_src_files
In order for the depmod command (now linked with kmod and thus running in
insmod_t) to handle System.map file (and other files residing in /usr/src and
needed during the "make modules_install" phase of a kernel build) we need it to
grant read rights to src_t.

Start with defining the interface.

See also bug #428322