Get grub2-install to work properly
The grub2-install application runs a few grub2-* commands in the background. Two
of those, grub2-bios-setup and grub2-probe, need access to the (fixed) disks, a
permission not granted to users by default (unless running in unconfined of
course).

Mark those two applications as bootloader_exec_t (as was the case with the
"grub" legacy command in the past) so these get their appropriate permissions
again.