Automatic filetrans to modules_object_t for /lib/modules
As per bug #428322, if /lib/modules does not exist, then "make modules_install"
will fail as it creates a directory (/lib/modules) with the wrong type (lib_t
instead of modules_object_t).

Add in a file transition when a directory named "modules" is created in a lib_t
location, transitioning to modules_object_t.

Had to add in the libs_lib_filetrans (from lib_t)
Had to add in the files_lib_filetrans_kernel_modules (to modules_object_t)