Support signed kernel modules
When using CONFIG_MODULE_SIG (signed kernel module support), when trying to load
a kernel module (such as "modprobe vxlan"), insmod_t wants to search through the
kernel keyring to find the key for the module. If it doesn't succeed, loading
fails:

modprobe: ERROR: could not insert 'vxlan': Required key not available.

In the dmesg output, we get:

Request for unknown module key '...' err -13

The denial shows that "insmod_t" wants to search on "kernel_t:key". Allowing
this is sufficient to allow signed modules to be loaded.