Support self-update of Portage
When Portage is asked to update itself, it triggers a specific procedure
(in portage/pym/portage/package/ebuild/doebuild.py, a method called
_prepare_self_update).

In this method, it will create a temporary copy of PORTAGE_BIN_PATH and
PORTAGE_PYM_PATH (which are bin_t, lib_t and some specific Portage
types).

During this copy, it needs to set the proper labels on the files, which
is why we now allow the various relabel operations.